Base support for brands/vendors

+++ src/BuildingBlocks/TinyCommerce.BuildingBlocks.Application/Queries/PagedResult.cs
@@ -5,7 +5,7 @@ namespace TinyCommerce.BuildingBlocks.Application.Queries
 {
     public class PagedResult<T> : List<T>
     {
-        public PagedResult(IEnumerable<T> items, int totalCount, int currentPage, int perPage)
+        public PagedResult(IEnumerable<T> items, long totalCount, int currentPage, int perPage)
         {
             Items = items;
             TotalCount = totalCount;
@@ -15,11 +15,17 @@ namespace TinyCommerce.BuildingBlocks.Application.Queries
         }
 
         public IEnumerable<T> Items { get; }
-        public int TotalCount { get; }
+        
+        public long TotalCount { get; }
+        
         public int CurrentPage { get; }
+        
         public int PerPage { get; }
+        
         public int TotalPages { get; }
+        
         public bool HasPreviousPage => (CurrentPage > 1);
+       
         public bool HasNextPage => (CurrentPage < TotalPages);
     }
 }
